+++ common/src/main/java/com/faboslav/structurify/common/config/data/structure/FlatnessCheckData.java
@@ -5,11 +5,18 @@ public final class FlatnessCheckData
 	public static final boolean OVERRIDE_GLOBAL_FLATNESS_CHECK_DEFAULT_VALUE = false;
 	public final static boolean IS_ENABLED_DEFAULT_VALUE = false;
 	public final static boolean ALLOW_NON_SOLID_BLOCKS_DEFAULT_VALUE = false;
+	public final static FlatnessCheckMode MODE_DEFAULT_VALUE = FlatnessCheckMode.AUTO;
+	public final static int MAX_HEIGHT_DIFFERENCE_DEFAULT_VALUE = 21;
+
+	public static final int MIN_HEIGHT_DIFFERENCE = 0;
+	public static final int MAX_HEIGHT_DIFFERENCE = 128;
 
 	private boolean overrideGlobalFlatnessCheck = OVERRIDE_GLOBAL_FLATNESS_CHECK_DEFAULT_VALUE;
 	private boolean defaultOverrideGlobalFlatnessCheck = OVERRIDE_GLOBAL_FLATNESS_CHECK_DEFAULT_VALUE;
 	private boolean isEnabled = IS_ENABLED_DEFAULT_VALUE;
 	private boolean defaultIsEnabled = IS_ENABLED_DEFAULT_VALUE;
+	private FlatnessCheckMode mode = MODE_DEFAULT_VALUE;
+	private int maxHeightDifference = MAX_HEIGHT_DIFFERENCE_DEFAULT_VALUE;
 	private boolean allowNonSolidBlocks = ALLOW_NON_SOLID_BLOCKS_DEFAULT_VALUE;
 
 	public FlatnessCheckData() {
@@ -58,7 +65,29 @@ public final class FlatnessCheckData
 		return this.allowNonSolidBlocks;
 	}
 
+	public FlatnessCheckMode getMode() {
+		return this.mode;
+	}
+
+	public void setMode(FlatnessCheckMode mode) {
+		this.mode = mode;
+	}
+
+	public int getMaxHeightDifference() {
+		return this.maxHeightDifference;
+	}
+
+	public void setMaxHeightDifference(int maxHeightDifference) {
+		this.maxHeightDifference = maxHeightDifference;
+	}
+
 	public void allowNonSolidBlocks(boolean allowNonSolidBlocks) {
 		this.allowNonSolidBlocks = allowNonSolidBlocks;
+	}
+
+	public enum FlatnessCheckMode
+	{
+		AUTO,
+		MANUAL
 	}
 }