+++ common/src/main/java/com/faboslav/structurify/common/config/serialization/FlatnessCheckDataSerializer.java
@@ -1,5 +1,6 @@
 package com.faboslav.structurify.common.config.serialization;
 
+import com.faboslav.structurify.common.config.data.structure.BiomeCheckData;
 import com.faboslav.structurify.common.config.data.structure.FlatnessCheckData;
 import com.google.gson.JsonObject;
 
@@ -8,6 +9,8 @@ public final class FlatnessCheckDataSerializer
 	private static final String OVERRIDE_GLOBAL_FLATNESS_CHECK_PROPERTY = "override_global_flatness_check";
 	private static final String ENABLE_FLATNESS_CHECK_PROPERTY = "enable_flatness_check";
 	private static final String FLATNESS_CHECK_ALLOW_NON_SOLID_PROPERTY = "flatness_check_allow_non_solid_blocks";
+	private static final String FLATNESS_CHECK_MODE_PROPERTY = "flatness_check_mode";
+	private static final String FLATNESS_CHECK_MAX_HEIGHT_DIFFERENCE = "flatness_check_max_height_difference";
 
 	public static void load(JsonObject structureJson, FlatnessCheckData flatnessCheckData) {
 		if (structureJson.has(OVERRIDE_GLOBAL_FLATNESS_CHECK_PROPERTY)) {
@@ -24,11 +27,23 @@ public final class FlatnessCheckDataSerializer
 			var allowNonSolidBlocks = structureJson.get(FLATNESS_CHECK_ALLOW_NON_SOLID_PROPERTY).getAsBoolean();
 			flatnessCheckData.allowNonSolidBlocks(allowNonSolidBlocks);
 		}
+
+		if (structureJson.has(FLATNESS_CHECK_MODE_PROPERTY)) {
+			var flatnessCheckMode = structureJson.get(FLATNESS_CHECK_MODE_PROPERTY).getAsString();
+			flatnessCheckData.setMode(FlatnessCheckData.FlatnessCheckMode.valueOf(flatnessCheckMode));
+		}
+
+		if (structureJson.has(FLATNESS_CHECK_MAX_HEIGHT_DIFFERENCE)) {
+			var maxHeightDifference = structureJson.get(FLATNESS_CHECK_MAX_HEIGHT_DIFFERENCE).getAsInt();
+			flatnessCheckData.setMaxHeightDifference(maxHeightDifference);
+		}
 	}
 
 	public static void save(JsonObject structureJson, FlatnessCheckData flatnessCheckData) {
 		structureJson.addProperty(OVERRIDE_GLOBAL_FLATNESS_CHECK_PROPERTY, flatnessCheckData.isOverridingGlobalFlatnessCheck());
 		structureJson.addProperty(ENABLE_FLATNESS_CHECK_PROPERTY, flatnessCheckData.isEnabled());
 		structureJson.addProperty(FLATNESS_CHECK_ALLOW_NON_SOLID_PROPERTY, flatnessCheckData.areNonSolidBlocksAllowed());
+		structureJson.addProperty(FLATNESS_CHECK_MODE_PROPERTY, flatnessCheckData.getMode().name());
+		structureJson.addProperty(FLATNESS_CHECK_MAX_HEIGHT_DIFFERENCE, flatnessCheckData.getMaxHeightDifference());
 	}
 }